requirements.txt
asgiref==3.7.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cloudinary==1.38.0
cryptography==41.0.7
defusedxml==0.7.1
dj-database-url==2.1.0
dj-rest-auth==5.0.2
Django==5.0.1
django-allauth==0.60.1
django-cloudinary-storage==0.3.0
django-cors-headers==4.3.1
djangorestframework==3.14.0
environs==10.3.0
gunicorn==21.2.0
idna==3.6
marshmallow==3.20.2
oauthlib==3.2.2
packaging==23.2
pillow==10.2.0
psycopg2-binary==2.9.9
pycparser==2.21
PyJWT==2.8.0
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
urllib3==2.1.0
whitenoise==6.6.0
